pVS: ----------------------------------------------------------------------
Purpose:
    Bug fix

Problem:
    In vg.h, the constant _HDF_VDATA was re-defined to -1.
    Similar to the tester problem in dist/hdf/test/tvattr.c, in hdp_vd.c
    there was a (int32)_HDF_VDATA as an argument to a function call.

Solution:
    Leave out the (int32) datatype changer.

Platform tested:
    Cray T90 - old floating point format

+++ mfhdf/dumper/hdp_vd.c
@@ -646,7 +646,7 @@ dumpvd_ascii(dump_info_t * dumpvd_opts,
                       if (dumpvd_opts->contents != DHEADER)
                         {
                             /* dump attributes */
-                            if (FAIL == dumpattr(vd_id, (int32)_HDF_VDATA, 1, ft, fp))
+                            if (FAIL == dumpattr(vd_id, _HDF_VDATA, 1, ft, fp))
                               {
                                   fprintf(stderr,"Failed to print vdata attributes for vd_id(%d) in file %s\n",
                                           (int) vd_id, file_name);
